# Conflicts:
#	.github/workflows/sync-repo-pr.yml.disabled
#	.github/workflows/sync-repo.yml.disabled
#	apps/api/plane/app/serializers/webhook.py
#	apps/api/plane/app/views/issue/attachment.py
#	apps/api/plane/bgtasks/work_item_link_task.py
#	apps/live/src/services/page/core.service.ts
#	pnpm-lock.yaml
#	pnpm-workspace.yaml

+++ apps/web/core/components/estimates/inputs/root.tsx
@@ -4,7 +4,6 @@
  * See the LICENSE file for details.
  */
 
-import type { FC } from "react";
 // plane imports
 import type { TEstimateSystemKeys } from "@plane/types";
 import { EEstimateSystem } from "@plane/types";
@@ -27,7 +26,7 @@ export function EstimateInputRoot(props: TEstimateInputRootProps) {
     case EEstimateSystem.POINTS:
       return (
         <EstimateNumberInput
-          value={value ? parseInt(value) : undefined}
+          value={value ? parseFloat(value) : undefined}
           handleEstimateInputValue={handleEstimateInputValue}
         />
       );
